backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==6.3.2
dataclasses-json==0.6.7
distro==1.9.0
Flask==2.3.3
Flask-Cors==3.0.10
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-text-splitters==0.3.2
langsmith==0.1.147
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.0
ollama==0.1.2
openai==1.57.1
orjson==3.10.12
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
py==1.11.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyPDF2==3.0.1
pytest==7.1.2
pytest-flask==1.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.3.7
yarl==1.18.3

# Base de datos y ORM
SQLAlchemy==2.0.36
alembic==1.13.1
psycopg2-binary==2.9.9

# Autenticación JWT
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0

# Cache y sesiones
redis==5.0.1
Flask-Session==0.5.0

# Validación y serialización
marshmallow==3.23.1
marshmallow-sqlalchemy==0.29.0

# Utilidades
python-dotenv==1.0.1
bcrypt==4.1.2
cryptography==42.0.5

# Testing adicional
pytest-cov==4.1.0
factory-boy==3.3.0

# Flask extensions
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1

# Dependencias para procesamiento de archivos y IA
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.12.0
Pillow==10.1.0
reportlab==4.0.4
